// src/libs/enums/product.enum.ts
export enum ProductSize {
  SMALL = "S",
  MEDIUM = "M",
  LARGE = "L",
  X_LARGE = "XL",
  TWO_XL = "2XL",
}

export enum ProductKidsSize {
  ONE_TWO = "1-2 years",
  THREE_FOUR = "3-4 years",
  FIVE_SIX = "5-6 years",
  SEVEN_EIGHT = "7-8 years",
  NINE_TEN = "9-10 years",
  ELEVEN_TWELVE = "11-12 years",
  THIRTEEN_FIFTEEN = "13-15 years",
}

export enum ProductShoeSize {
  THIRTY_EIGHT = 38,
  THIRTY_NINE = 39,
  FORTY = 40,
  FORTY_ONE = 41,
  FORTY_TWO = 42,
  FORTY_THREE = 43,
  FORTY_FOUR = 44,
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  MEN = "MEN",
  WOMEN = "WOMEN",
  KIDS = "KIDS",
  SHOES = "SHOES",
}